education/requirement.txt
certifi==2018.1.18
chardet==3.0.4
dj-static==0.0.6
Django==2.0.3
django-braces==1.12.0
django-embed-video==1.1.2
django-json-response==1.1.5
django-memcache-status==1.3
django-memcached==0.1.2
djangorestframework==3.7.7
idna==2.6
Pillow==5.0.0
python-memcached==1.59
python-memcached-stats==0.1
python3-memcached==1.51
pytz==2018.3
requests==2.18.4
six==1.11.0
static3==0.7.0
urllib3==1.22
